ci: analyze with clang-tidy and test under sanitizers
The Windows port will meet MSVC's stricter narrowing rules with no
local toolchain to try them first; clang-tidy's narrowing checks catch
the same class of bug on every push, alongside the bugprone and
performance sets. The configuration is curated for signal - checks
that flag every coordinate pair, Catch2's static test registration, or
deliberate pragma once stay off, and tests additionally drop the
unchecked-optional check that cannot see through REQUIRE.

An opt-in SIDESCOPES_SANITIZE build flag feeds two new CI jobs:
address+undefined and thread. The analysis worker hands frames to the
UI through a deliberately lock-free read; that design now has a race
detector auditing it on every push instead of a comment asserting it
safe.

The findings the new analysis surfaced are fixed here: implicit
int-to-float and widening-multiplication conversions in the scope
engines and tests, and preferences parsing moved from atoi to strtol.

+++ src/core/marker_smoother.cpp
@@ -15,9 +15,9 @@ FloatColor AverageNeighborhood(const FrameView& frame, int px, int py, int radiu
             if (sample_x < 0 || sample_x >= frame.width || sample_y < 0 || sample_y >= frame.height)
                 continue;
             const Color color = frame.ColorAt(sample_x, sample_y);
-            sum_r += color.r;
-            sum_g += color.g;
-            sum_b += color.b;
+            sum_r += static_cast<float>(color.r);
+            sum_g += static_cast<float>(color.g);
+            sum_b += static_cast<float>(color.b);
             ++count;
         }
     }
